[faces] set all/only `hl-todo` keywords which are present
Instead of hard-coding the value of `hl-todo-keyword-faces`, alter the
value of any keyword in the list for which we have a value in the
theme's list. This allows users to alter the list before loading the
theme (and retain their own additions) while still using the theme
colours where relevant.